+++ clients/ClientMySQL.ts
@@ -12,6 +12,7 @@ import {
   MAX_FILE_NAME_LENGTH,
   TABLE_MIGRATIONS,
 } from "../consts.ts";
+import { NessieError } from "../cli/errors.ts";
 
 export type { MySQLClientOptions };
 
@@ -87,7 +88,7 @@ export class ClientMySQL extends AbstractClient<MySQLClient> {
         if (e?.message === "Query was empty") {
           ra.push(undefined);
         } else {
-          throw new Error(query + "\n" + e + "\n" + ra.join("\n"));
+          throw new NessieError(query + "\n" + e + "\n" + ra.join("\n"));
         }
       }
     }