+++ mod.ts
@@ -7,3 +7,4 @@ export * from "./wrappers/AbstractMigration.ts";
 export * from "./wrappers/AbstractSeed.ts";
 export { MAX_FILE_NAME_LENGTH, REGEXP_MIGRATION_FILE_NAME } from "./consts.ts";
 export { isMigrationFile } from "./cli/utils.ts";
+export { NessieError } from "./cli/errors.ts";